+++ apps/web-antd/src/views/monitor/online/index.vue
@@ -1,12 +1,17 @@
 <script setup lang="ts">
-import type { Recordable } from '@vben/types';
+import type { VbenFormProps } from '@vben/common-ui';
 
-import { Page, type VbenFormProps } from '@vben/common-ui';
+import type { VxeGridProps } from '#/adapter/vxe-table';
+import type { OnlineUser } from '#/api/monitor/online/model';
+
+import { ref } from 'vue';
+
+import { Page } from '@vben/common-ui';
 import { getVxePopupContainer } from '@vben/utils';
 
 import { Popconfirm } from 'ant-design-vue';
 
-import { useVbenVxeGrid, type VxeGridProps } from '#/adapter/vxe-table';
+import { useVbenVxeGrid } from '#/adapter/vxe-table';
 import { forceLogout, onlineList } from '#/api/monitor/online';
 
 import { columns, querySchema } from './data';
@@ -22,6 +27,7 @@ const formOptions: VbenFormProps = {
   wrapperClass: 'grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4',
 };
 
+const onlineCount = ref(0);
 const gridOptions: VxeGridProps = {
   columns,
   height: 'auto',
@@ -31,10 +37,12 @@ const gridOptions: VxeGridProps = {
   },
   proxyConfig: {
     ajax: {
-      query: async (_, formValues) => {
-        return await onlineList({
+      query: async (_, formValues = {}) => {
+        const resp = await onlineList({
           ...formValues,
         });
+        onlineCount.value = resp.total;
+        return resp;
       },
     },
   },
@@ -51,13 +59,9 @@ const gridOptions: VxeGridProps = {
 
 const [BasicTable, tableApi] = useVbenVxeGrid({ formOptions, gridOptions });
 
-async function handleForceOffline(row: Recordable<any>) {
+async function handleForceOffline(row: OnlineUser) {
   await forceLogout(row.tokenId);
   await tableApi.query();
-}
-
-function onlineCount() {
-  return tableApi?.grid?.getData?.()?.length ?? 0;
 }
 </script>
 
@@ -68,7 +72,7 @@ function onlineCount() {
         <div class="mr-1 pl-1 text-[1rem]">
           <div>
             在线用户列表 (共
-            <span class="text-primary font-bold">{{ onlineCount() }}</span>
+            <span class="text-primary font-bold">{{ onlineCount }}</span>
             人在线)
           </div>
         </div>