+++ apps/web-antd/src/views/system/oss/index.vue
@@ -1,10 +1,14 @@
 <script setup lang="ts">
-import type { Recordable } from '@vben/types';
+import type { VbenFormProps } from '@vben/common-ui';
+
+import type { VxeGridProps } from '#/adapter/vxe-table';
+import type { PageQuery } from '#/api/common';
+import type { OssFile } from '#/api/system/oss/model';
 
 import { onMounted, ref } from 'vue';
 import { useRouter } from 'vue-router';
 
-import { Page, useVbenModal, type VbenFormProps } from '@vben/common-ui';
+import { Page, useVbenModal } from '@vben/common-ui';
 import { $t } from '@vben/locales';
 import { getVxePopupContainer } from '@vben/utils';
 
@@ -14,6 +18,7 @@ import {
   Modal,
   Popconfirm,
   Space,
+  Spin,
   Switch,
   Tooltip,
 } from 'ant-design-vue';
@@ -22,13 +27,12 @@ import {
   addSortParams,
   useVbenVxeGrid,
   vxeCheckboxChecked,
-  type VxeGridProps,
 } from '#/adapter/vxe-table';
 import { configInfoByKey } from '#/api/system/config';
 import { ossDownload, ossList, ossRemove } from '#/api/system/oss';
 import { downloadByData } from '#/utils/file/download';
 
-import { columns, querySchema } from './data';
+import { columns, fallbackImageBase64, querySchema } from './data';
 import fileUploadModal from './file-upload-modal.vue';
 import imageUploadModal from './image-upload-modal.vue';
 
@@ -67,7 +71,7 @@ const gridOptions: VxeGridProps = {
   proxyConfig: {
     ajax: {
       query: async ({ page, sorts }, formValues = {}) => {
-        const params: any = {
+        const params: PageQuery = {
           pageNum: page.currentPage,
           pageSize: page.pageSize,
           ...formValues,
@@ -101,7 +105,7 @@ const [BasicTable, tableApi] = useVbenVxeGrid({
   },
 });
 
-async function handleDownload(row: Recordable<any>) {
+async function handleDownload(row: OssFile) {
   const hideLoading = message.loading($t('pages.common.downloadLoading'), 0);
   try {
     const data = await ossDownload(row.ossId);
@@ -111,14 +115,14 @@ async function handleDownload(row: Recordable<any>) {
   }
 }
 
-async function handleDelete(row: Recordable<any>) {
-  await ossRemove(row.ossId);
+async function handleDelete(row: OssFile) {
+  await ossRemove([row.ossId]);
   await tableApi.query();
 }
 
 function handleMultiDelete() {
   const rows = tableApi.grid.getCheckboxRecords();
-  const ids = rows.map((row: any) => row.ossId);
+  const ids = rows.map((row: OssFile) => row.ossId);
   Modal.confirm({
     title: '提示',
     okType: 'danger',
@@ -192,11 +196,22 @@ const [FileUploadModal, fileUploadApi] = useVbenModal({
         </Space>
       </template>
       <template #url="{ row }">
+        <!-- placeholder为图片未加载时显示的占位图 -->
+        <!-- fallback为图片加载失败时显示 -->
+        <!-- 需要设置key属性 否则切换翻页会有延迟 -->
         <Image
+          :key="row.ossId"
           v-if="preview && isImageFile(row.url)"
           :src="row.url"
           height="50px"
-        />
+          :fallback="fallbackImageBase64"
+        >
+          <template #placeholder>
+            <div class="flex size-full items-center justify-center">
+              <Spin />
+            </div>
+          </template>
+        </Image>
         <span v-else>{{ row.url }}</span>
       </template>
       <template #action="{ row }">